Conflicts:
	src/main/java/edu/ncsu/csc573/project/common/schema/ACKType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ACKTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ChangePasswordParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ChangePasswordResponseParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ChangePasswordResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ChangePasswordType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/CommandRequestType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/CommandResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/DownloadFileParamType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/FileParamType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ForgotPWDResponseParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ForgotPWDResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ForgotPasswdParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ForgotPasswdType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/GetResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/GetResponseTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/GetType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/GetTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/InvalidResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/InvalidResponseTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/JoinResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/JoinResponseTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/JoinType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/JoinTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LeaveResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LeaveResponseTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LeaveType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LeaveTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LoginParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LoginResponseParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LoginResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LoginType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LogoutParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LogoutResponseParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LogoutResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/LogoutType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/MatchFileParamType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/ObjectFactory.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PublishResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PublishResponseTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PublishType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PublishTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PutResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PutResponseTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PutType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/PutTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/RegisterParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/RegisterResponseParamsType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/RegisterResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/RegisterType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/Request.java
	src/main/java/edu/ncsu/csc573/project/common/schema/Response.java
	src/main/java/edu/ncsu/csc573/project/common/schema/SearchResponseType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/SearchResponseTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/SearchType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/SearchTypeParams.java
	src/main/java/edu/ncsu/csc573/project/common/schema/TableParamType.java
	src/main/java/edu/ncsu/csc573/project/common/schema/package-info.java

+++ src/main/java/edu/ncsu/csc573/project/controllayer/hashspacemanagement/IQuery.java
@@ -20,5 +20,5 @@ public interface IQuery {
 	 * This 
 	 * @return
 	 */
-	public byte[] getCordinates();
+	public int[] getCordinates();
 }
